Multiple Sparse PCA

Returns multiple sparse principal component of a matrix using an iterative deflation heuristic.

mspca(
  Sigma,
  r,
  ks,
  feasibilityConstraintType = 0L,
  verbose = TRUE,
  maxIter = 200L,
  feasibilityTolerance = 1e-04,
  stallingTolerance = 1e-08,
  timeLimitTPM = 20L,
  maxRestartTPM = 20L,
  minRestartTPM = 10L
)

Arguments

Sigma

A matrix. The correlation or covariance matrix, whose sparse PCs will be computed.

r

An integer. Number of principal components (PCs) to be computed.

ks

A list of integers. Target sparsity of each PC.

feasibilityConstraintType

(optional) An integer. Type of feasibility constraints to be enforced. 0: orthogonality constraints; 1: uncorrelatedness constraints. Default 0.

verbose

(optional) A Boolean. Controls console output. Default TRUE.

maxIter

(optional) An integer. Maximum number of iterations of the algorithm. Default 200.

feasibilityTolerance

(optional) A float. Tolerance for the violation of the orthogonality constraints. Default 1e-4

stallingTolerance

(optional) A float. Controls the objective improvement below which the algorithm is considered to have stalled. Default 1e-8

timeLimitTPM

(optional) An integer. Maximum time in seconds for the truncated power method (inner iteration). Default 20.

maxRestartTPM

(optional) An integer. Number of random restarts of the truncated power method (inner iteration) for the first outer iteration. Default 20.

minRestartTPM

(optional) An integer. Number of random restarts of the truncated power method (inner iteration) for outer iterations >= 2. Default 10.

Value

An object with 4 fields: `x_best` (p x r array containing the sparse PCs), `objective_value`, `feasibility_violation`, `runtime`.

Examples

TestMat <- cor(datasets::mtcars)
mspca(TestMat, 2, c(4,4))
#> ---- Iterative deflation algorithm for sparse PCA with multiple PCs ---
#> Dimension: 11
#> Number of PCs: 2
#> Sparsity pattern:  4 4
#>   Iteration |      Objective value |     Feasibility Violation |       Time
#>           1 |                0.649 |                  1.00e+00 |      0.000
#>          11 |                0.642 |                  7.33e-01 |      0.001
#>          21 |                0.641 |                  7.11e-01 |      0.002
#>          31 |                0.614 |                  2.12e-01 |      0.004
#>          36 |                0.597 |                  0.00e+00 |      0.006
#> $objective_value
#> [1] 6.568963
#> 
#> $feasibility_violation
#> [1] 0
#> 
#> $runtime
#> [1] 0.006
#> 
#> $x_best
#>             [,1]       [,2]
#>  [1,] -0.5072433  0.0000000
#>  [2,]  0.0000000  0.5061949
#>  [3,]  0.5146686  0.0000000
#>  [4,]  0.0000000  0.5095740
#>  [5,] -0.4575700  0.0000000
#>  [6,]  0.5181217  0.0000000
#>  [7,]  0.0000000 -0.4714724
#>  [8,]  0.0000000 -0.5116785
#>  [9,]  0.0000000  0.0000000
#> [10,]  0.0000000  0.0000000
#> [11,]  0.0000000  0.0000000
#>